+++ app/recommenders/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import RecommenderWorkflowDashboard from "@/components/recommenders/RecommenderWorkflowDashboard";
+
+export default function RecommendersPage() {
+  return <RecommenderWorkflowDashboard />;
+}