+++ lib/opportunity-toolkit/index.ts
@@ -3,3 +3,5 @@ export * from "./recommendationLetterStudio";
 export * from "./bragSheet";
 export * from "./portfolioExport";
 export * from "./applicationAssistant";
+
+export * from "./pdf/portfolioPdf";